refactor: Move shared save/load logic to Storage base class
Refactored `Storage.save` and `Storage.load` to follow the template method pattern,
similar to `evict`. The base class now handles default key generation and
short-hand digest expansion, while subclasses implement `_save` and `_load`.

This removes boilerplate from `Memory`, `PickleFile`, `CloudpickleFile`,
`BagOfHoldingH5File`, and `Sql` storage backends.
