+++ src/fleche/storage/cloudpickle_file.py
@@ -6,7 +6,7 @@ from typing import Any
 from cloudpickle import loads, dumps
 
 from .file import FileStorage
-from ..digest import digest, Digest, DIGEST_LENGTH
+from ..digest import Digest
 
 
 @dataclass
@@ -15,15 +15,11 @@ class CloudpickleFile(FileStorage):
     Store values as files on the filesystem using cloudpickle for serialization.
     """
 
-    def save(self, value: Any, key: Digest | None = None) -> str:
-        if key is None:
-            key = digest(value)
+    def _save(self, value: Any, key: Digest) -> str:
         (self._path(key)).write_bytes(dumps(value))
         return key
 
-    def load(self, key: str) -> Any:
-        if len(key) < DIGEST_LENGTH:
-            key = self.expand(key)
+    def _load(self, key: str) -> Any:
         try:
             return loads((self._path(key)).read_bytes())
         except FileNotFoundError:
