+++ src/fleche/storage/pickle_file.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from typing import Any
 
 from .file import FileStorage
-from ..digest import digest, Digest, DIGEST_LENGTH
+from ..digest import Digest
 
 
 @dataclass
@@ -14,15 +14,11 @@ class PickleFile(FileStorage):
     Store values as files on the filesystem using the standard pickle module for serialization.
     """
 
-    def save(self, value: Any, key: Digest | None = None) -> str:
-        if key is None:
-            key = digest(value)
+    def _save(self, value: Any, key: Digest) -> str:
         (self._path(key)).write_bytes(pickle.dumps(value))
         return key
 
-    def load(self, key: str) -> Any:
-        if len(key) < DIGEST_LENGTH:
-            key = self.expand(key)
+    def _load(self, key: str) -> Any:
         try:
             return pickle.loads((self._path(key)).read_bytes())
         except FileNotFoundError:
